% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{as_sstable}
\alias{as_sstable}
\alias{as_sstable.default}
\alias{as_sstable.list}
\alias{as_sstable.logist_summary}
\title{Coerce an object to sstable}
\usage{
as_sstable(x, ...)

\method{as_sstable}{default}(x, flextable = FALSE, ...)

\method{as_sstable}{list}(x, flextable = FALSE, ...)

\method{as_sstable}{logist_summary}(x, include_footnote = TRUE, flextable = FALSE, ...)
}
\arguments{
\item{x}{An object, usually a named list of length 2 whose names are 'table' and 'footer', or a data.frame/matrix (optionally with attribute "footer")}

\item{...}{additional param passed to \link{ss_flextable}}

\item{flextable}{logical value specifying whether to return a flextable. Default is FALSE}

\item{include_footnote}{logical value specifying whether to include footnote in the output. Default is FALSE}
}
\value{
A matrix of class ss_tbl if flextable == FALSE or a flextable
}
\description{
A function to coerce objects to a sstable
}
