% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstable.R
\name{contSummary}
\alias{contSummary}
\title{Calculate summary statistics for continuous variable}
\usage{
contSummary(x, statistics = c("med.IQR", "med.90", "med.range",
  "mean.sd"), digits = 1, n = TRUE)
}
\arguments{
\item{x}{a numeric vector to be summarised.}

\item{statistics}{a character specifies summary statistics for continuous row variables.}

\item{digits}{a number specifies number of significant digits for numeric statistics.}

\item{n}{a logical value specifies whether output will contain the number of non-missing values.}
}
\value{
a string displays formatted summary statistics.
}
\description{
A function to calculate summary statistics for continuous variable.
}
