% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUCRU_data_dictionary.R
\name{convert.info}
\alias{convert.info}
\title{Convert imported OUCRU's data dictionary into a general data dictionary}
\usage{
convert.info(oucru_info, oucru_category)
}
\arguments{
\item{oucru_info}{a data frame contains imported data of a sheet from OUCRU's data dictionary.}

\item{oucru_category}{a data frame specifies how categorical variables are coded in OUCRU's database.}
}
\value{
A data frame contains data dictionary information in a general format.
}
\description{
A function to convert imported OUCRU's data dictionary into a general data dictionary.
}
