% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{Create project's folder based on pre-defined folder structure}
\usage{
createProject(dir, structure = c("OUCRU_RCT_wInterim"))
}
\arguments{
\item{dir}{a character specifies where to create the project's folder.}

\item{structure}{a character specifies the pre-defined folder structure to be used.}
}
\value{
A message to confirm that the project's folder is created at the designed place.
}
\description{
A generic function to create a project's folder based on pre-defined folder structure.
}
