% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_summary.R
\name{explicit.lm}
\alias{explicit.lm}
\title{Transform a model to its explicit form}
\usage{
\method{explicit}{lm}(fit, data = NULL)
}
\arguments{
\item{fit}{An object of class "lm"}
}
\value{
A model with explicit formula
}
\description{
A function to transform an lm model formula to its explicit form.
This is useful when you want to use stats::drop1 to drop additional terms in interaction models.
}
\examples{
iris.model <- lm(Sepal.Width ~ Sepal.Length*Petal.Width, data=iris)
explicit(iris.model)
}
\seealso{
\link[stats]{add1}
}
