% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit.R
\name{explicit}
\alias{explicit}
\title{Transform an object to its explicit form}
\usage{
explicit(x, ...)
}
\arguments{
\item{x}{An object}
}
\description{
A function to transform an object to its explicit form if possible
}
\seealso{
\link{explicit.lm}
}
