% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlabel.R
\name{getlabel.default}
\alias{getlabel.default}
\title{Functions to get variable label}
\usage{
getlabel.default(x, units = TRUE, fit = FALSE)
}
\arguments{
\item{x}{Variable or data to get label}

\item{units}{Append unit in the label?}

\item{fit}{Use presentation as in model fitting?}

\item{meta}{Metadata of the dataset}
}
\value{
vector of character of label
}
\description{
Functions to get variable label
}
