% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlabel.R
\name{getlabel}
\alias{getlabel}
\title{Obtain label of variable(s) based on metadata information}
\usage{
getlabel(x, meta = NULL, unit = TRUE, fit = FALSE)
}
\arguments{
\item{x}{a vector or data frame to get label.}

\item{meta}{a data frame contains metadata of the dataset. This data frame should have the following variables: varname (character variable specifies name of each variable), label (character variable specifies label of each variables), type (character variable specifies type of each variable [numeric, factor, character, datetime], unit (character variable specifies units of each continuous variable), scale (numeric variable specifies how each variable will be scaled).}

\item{unit}{a logical value specifies whether unit will be appendded in the label of continous variables.}

\item{fit}{a logical value specifies whether the label will use presentation as in model fitting.}
}
\value{
vector of character of label
}
\description{
A function to obtain label of one vector or each variable in one data frame based on pre-defined metadata information.
}
