% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstable_baseline.R
\name{getxs}
\alias{getxs}
\title{Summarise dataset}
\usage{
getxs(xformula)
}
\arguments{
\item{formula}{Specification of dependent and independent variables}

\item{data}{Data to summarise}

\item{bycol}{Summary by column or by row?}

\item{pooledGroup}{Pooling all subgroups of dependent variable?}

\item{statistics}{Summary statistics for continuous variables?}

\item{continuous}{Independent variables are continuous variables?}

\item{test}{Display p values from statistical test comparing independent variables between sub-group of dependent variable?}

\item{digits}{Number of decimal digit number}

\item{pdigits}{Number of decimal digit number for p values}

\item{pcutoff}{Cut-off to display p values}

\item{chisq.test}{Use Chi-squared test rather than Fisher's exact test for categorical variables}

\item{correct}{Correction for Chi-squared test}

\item{workspace}{Parameter for Fisher's exact test}

\item{hybrid}{Parameter for Fisher's exact test}

\item{simulate.p.value}{Parameter for Fisher's exact test}

\item{B}{Parameter for Fisher's exact test}

\item{footer}{Footer(s) for table}

\item{flextable}{Create output as a flextable-type table}

\item{bg}{Background color for odd rows of table}
}
\value{
a flextable-type table or a list with values/headers/footers
}
\description{
Summarise dataset
}
