% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ajsurvplot.R
\name{gg_ajsurvplot2}
\alias{gg_ajsurvplot2}
\title{Plot 2 Aalen-Johansen curves with ggplot for competing risks}
\usage{
gg_ajsurvplot2(
  formula,
  data,
  weights,
  subset,
  na.action,
  main.event,
  competing.event,
  facet.by = ~strata,
  count,
  id,
  timefix,
  ...,
  .return_data = FALSE
)
}
\arguments{
\item{formula}{A standard model formula, with survival on the left and covariates on the right}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model}

\item{weights, subset, na.action, , count, id, timefix}{parameter passed to finegray model. See \code{\link[survival:finegray]{finegray}},}

\item{main.event, competing.event}{main and competing event to plot}

\item{facet.by}{formula. default to all strata}

\item{...}{optional parameters passed to \code{ggsurvplot}.}

\item{.return_data}{[FALSE] if TRUE, get the data instead.}
}
\description{
Extend \link[ggsurvfit:ggsurvfit]{ggsurvfit} functionality to plot competing risk curve, with main risk and competing risk in one plot
}
