% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_boxcox.R
\name{gg_boxcox}
\alias{gg_boxcox}
\title{Compute and draw Box-Cox transformation plot for linear models using ggplot2}
\usage{
gg_boxcox(object, ...)
}
\arguments{
\item{object}{a formula or fitted model object. Currently only lm and aov objects are handled.}

\item{...}{additional parameters to be used in the model fitting.}

\item{lambda}{vector of values of lambda – default (-2, 2) in steps of 0.1}

\item{interp}{logical which controls whether spline interpolation is used. Default to TRUE if plotting with lambda of length less than 100.}

\item{eps}{Tolerance for lambda = 0; defaults to 0.02}

\item{xlab}{defaults to "lambda"}

\item{ylab}{defaults to "log-Likelihood"}
}
\value{
a ggplot object
}
\description{
A forked version of MASS::boxcox using ggplot2
}
\examples{
gg_boxcox(Volume ~ log(Height) + log(Girth), data = trees,
          lambda = seq(-0.25, 0.25, length = 10))
}
\seealso{
\link[MASS]{boxcox}
}
