% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{ht_theme_kable}
\alias{ht_theme_kable}
\title{A kable-esque theme for huxtable object}
\usage{
ht_theme_kable(
  ht,
  header_rows = 1:2,
  header_cols = NULL,
  border_width = 1,
  border_color = "#dddddd",
  bg = c("white", "#f9f9f9")
)
}
\arguments{
\item{ht}{an object of class huxtable}

\item{header_rows}{a numeric vector that delimits the header zone.}

\item{border_width}{a number that defines huxtable border width}

\item{border_color}{a character string that defines huxtable border color}

\item{bg}{a character vector that defines background color of the flextable. If length(bg) >= 2, the table will have stripe background, otherwise plain.}
}
\value{
an object of class huxtable
}
\description{
This function provides a kable-esque theme for huxtable object
}
