% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{ht_theme_markdown}
\alias{ht_theme_markdown}
\title{A stripe theme for huxtable object}
\usage{
ht_theme_markdown(ht, header_rows = 1:2, header_cols = NULL,
  border_width = 0.8, border_color = grey(0.75), bg = c(grey(0.95),
  "white"))
}
\arguments{
\item{ht}{an object of class huxtable}

\item{header_rows}{a numeric vector that delimits the header zone.}

\item{border_width}{a number that defines huxtable border width}

\item{border_color}{a character string that defines huxtable border color}

\item{bg}{a character vector that defines background color of the flextable. If length(bg) >= 2, the table will have stripe background, otherwise plain.}
}
\value{
an object of class huxtable
}
\description{
This function provides a stripped theme for huxtable object
}
