% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUCRU_data_dictionary.R
\name{import.info}
\alias{import.info}
\title{Import OUCRU's data dictionary into R}
\usage{
import.info(table_name, input, output)
}
\arguments{
\item{table_name}{a character vector specifies names of Excel sheets (in OUCRU's data dictionary) to be imported.}

\item{input}{a character value specifies path to the OUCRU's data dictionary file (in Excel format).}

\item{output}{a character value specifies where to save imported data.}
}
\value{
Nothing (imported data will be saved in the pre-defined output directory)
}
\description{
A function to import OUCRU's data dictionary into R.
}
