% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labAE.R
\name{labAE}
\alias{labAE}
\title{Identify new lab adverse event to report}
\usage{
labAE(data, gradecut)
}
\arguments{
\item{data}{a numeric vector of consecutive lab adverse event grade.}

\item{gradecut}{a grade threshold to report.}
}
\value{
A factor specifies whether new lab adverse event to report is identified or not.
}
\description{
A function to identify new lab adverse event to report from a vector of consecutive lab adverse event grade.
}
