% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_summary.R
\name{logist_summary}
\alias{logist_summary}
\title{An alternative summary for a logistic model with OR, CI, and p-values}
\usage{
logist_summary(
  fit,
  method = c("lik.ratio", "wald"),
  stat_digits = 2,
  p_digits = 4,
  verbose = FALSE
)
}
\arguments{
\item{fit}{a logistic regression model of class "glm"}

\item{method}{a string whose value is either
"lik.ratio" for CI and tests based on likelihood ratio statistics (preferred)
or "wald" for CI and tests based on Wald statistics}

\item{stat_digits}{Number of decimal digits for statistics}

\item{p_digits}{Number of decimal digits for the p-values}
}
\value{
A data.frame of additional class "logist_summary"
}
\description{
a function to create summary table for glm logistic model
}
\author{
Marcel Wolbers, Lam Phung Khanh, and Trinh Dong Huu Khanh
}
