% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_mult.R
\name{mutate_f}
\alias{mutate_f}
\alias{mutate_f.data.table}
\alias{mutate_f.tbl}
\alias{mutate_f.grouped_df}
\alias{mutate_f.data.frame}
\title{Create or transform variables, with support for multiple assignment}
\usage{
mutate_f(.data, ...)

\method{mutate_f}{data.table}(.data, ...)

\method{mutate_f}{tbl}(.data, ...)

\method{mutate_f}{grouped_df}(.data, ...)

\method{mutate_f}{data.frame}(.data, ...)
}
\arguments{
\item{.data}{A data.frame, data.table, or tbl}

\item{...}{Name-value pairs of expressions, following data.table syntax: "new_col" := expression.
Multiple assignments can be done via c("new_col1", "new_col2") := expression_returning_a_list.

If unnamed, names of each list elements will be used}
}
\value{
a data.frame, data.table, or a tbl
}
\description{
A function to create new variable(s) based on existing ones.
}
\examples{
df <- data.table(baz = 1:10, foo = c(rep(1, 5), rep(2, 5)))

df \%>\%
dplyr::group_by(foo) \%>\%
mutate_f(c('foo', 'bar') := list(a = cumsum(baz), b = baz*2))
}
