% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_join.R
\name{overlap_join}
\alias{overlap_join}
\title{Overlap join of 2 data frames}
\usage{
overlap_join(
  x,
  y,
  by = NULL,
  by.x = by,
  by.y = by,
  type = c("left", "right", "inner"),
  overlap = c("any", "within", "start", "end", "equal"),
  multiple_match = c("all", "first", "last"),
  max_gap = 0L,
  min_overlap = 1L
)
}
\arguments{
\item{x, y}{data.frames}

\item{by, by.x, by.y}{a character vector of variables to join by, passed to \link[data.table]{foverlaps}' by.x ad by.y.

If NULL, the default,  will do a natural join, using all variables with common names across the two tables.
A message lists the variables so that you can check they're right
(to suppress the message, simply explicitly list the variables that you want to join).

Otherwise, the last two columns must be the start and end time of the event specified.}

\item{type}{character-typed value, either 'left' (default), right, or 'inner',
representing left_join, right_join, and inner_join}

\item{overlap}{character-typed value, passed to \link[data.table]{foverlaps}' type

Allowed values are 'any' (default), 'within', 'start', 'end' and 'equal'.

The types shown here are identical in functionality
to the function findOverlaps in the bioconductor package IRanges.

Let [a,b] and [c,d] be intervals in x and y with a<=b and c<=d.
For type="start", the intervals overlap iff a == c.
For type="end", the intervals overlap iff b == d.
For type="within", the intervals overlap iff a>=c and b<=d.
For type="equal", the intervals overlap iff a==c and b==d.
For type="any", as long as c<=b and d>=a, they overlap.
In addition to these requirements, they also have to satisfy the min_overlap argument.}

\item{multiple_match}{a character-typed value, passed to \link[data.table]{foverlaps}' mult

When multiple rows in y match to the row in x,
multiple_match controls which values are returned - "all" (default), "first" or "last".}

\item{max_gap}{a character-typed value, passed to \link[data.table]{foverlaps}' maxgap.
It should be a non-negative integer value, >= 0. Default is 0 (no gap). For intervals [a,b] and [c,d],
where a<=b and c<=d, when c > b or d < a, the two intervals don't overlap.
If the gap between these two intervals is <= maxgap, these two intervals are considered as overlapping.}

\item{min_overlap}{a character-typed value, passed to \link[data.table]{foverlaps}' minoverlap

It should be a positive integer value, > 0. Default is 1. For intervals [a,b] and [c,d],
where a<=b and c<=d, when c<=b and d>=a, the two intervals overlap.
If the length of overlap between these two intervals is >= minoverlap,
then these two intervals are considered to be overlapping.}
}
\value{
a tibble with relevant columns from x, y
}
\description{
This function is a wrapper of data.table's foverlaps, following the dplyr's _join grammar.
}
\seealso{
\link[data.table]{foverlaps}
}
