% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_join.R
\name{overlap_summary}
\alias{overlap_summary}
\title{Summarise events overlapping of 2 event data frames.}
\usage{
overlap_summary(
  x,
  y,
  ids,
  cols,
  cols.x = cols,
  cols.y = cols,
  sstable = FALSE,
  flextable = FALSE
)
}
\arguments{
\item{x, y}{data.frames.}

\item{ids}{the id col(s) wrapped in dplyr::vars() distinguishing different events.
Based on this is the summarisation taken.}

\item{cols, cols.x, cols.y}{variables wrapped in dplyr::vars(), passed to overlap_join.
By default, except for the last 2, every columns will be automatically included into ids.}

\item{sstable}{logical value specifying whether to return in sstable format. Default is FALSE.}

\item{flextable}{logical value specifying whether to build flextable object. Default it FALSE. Set to TRUE forces sstable to TRUE.}
}
\value{
A summary table that has at least 6 columns.

id column(s): identification of each event.

X: total time when X happens, regardless Y
Y: total time when Y happens, regardless X
X_Y: time duration when X and Y happen simultaneously
X_notY: time duration when only X happens
notX_Y: time duration when only Y happens
}
\description{
A function to summarise the overlapping status of 2 event data frames
}
\seealso{
\link[dplyr]{vars}
}
