% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkits.R
\name{pct}
\alias{pct}
\title{Get the proportion for each level or range of a vector}
\usage{
pct(x, method = c("auto", "bin", "cont", "fct"), na.rm = TRUE, ...)
}
\arguments{
\item{x}{An object}

\item{method}{A method to calculate the percentage.
Default is 'auto' which will attempt to choose the most fit method.
'bin' is only meaningful for vector that only has to values.
'fct' is meaningful for descrete variable
'cont' is used for continuous variables. This will break the variable into several ranges
depended on the breaks specified in ...}

\item{na.rm}{A logical value deciding whether NA should be removed}

\item{...}{Break points to cut the continuous variables into ranges. These are passed to function cut.
See \link{cut}}
}
\value{
If method == 'bin': the percentage of non-reference level

If method == 'fct': the percentages of all levels

If method == 'cont': the percentages of all ranges based on specified breaks
}
\description{
A function to calculate the proportion of each value range
}
