% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_summary.R
\name{print.logist_summary}
\alias{print.logist_summary}
\title{Print method for logist_summary}
\usage{
\method{print}{logist_summary}(x, ...)
}
\arguments{
\item{x}{an object of class logist_summary}

\item{...}{additional params passed to print.data.frame}
}
\value{
invisibly return itself
}
\description{
Print method for logist_summary table
}
\seealso{
\link{print.data.frame}
}
