% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_effect.R
\name{print.subgroup_logist_summary}
\alias{print.subgroup_logist_summary}
\title{Print method for subgroup_logist_summary}
\usage{
\method{print}{subgroup_logist_summary}(x, ...)
}
\arguments{
\item{x}{an object of class subgroup_logist_summary}

\item{...}{additional params passed to print.data.frame}
}
\value{
invisibly return itself
}
\description{
Print method for subgroup_logist_summary
}
\seealso{
\link{print.data.frame}
}
