% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{rbind.ss_tbl}
\alias{rbind.ss_tbl}
\alias{rbind.ae_tbl}
\alias{rbind.survcomp_tbl}
\alias{rbind.baseline_tbl}
\title{Row-binding two sstables}
\usage{
\method{rbind}{ss_tbl}(tbl1, tbl2, footer = NULL, ...)

\method{rbind}{ae_tbl}(tbl1, tbl2, footer = NULL)

\method{rbind}{survcomp_tbl}(tbl1, tbl2, footer = NULL)

\method{rbind}{baseline_tbl}(tbl1, tbl2, footer = NULL)
}
\arguments{
\item{tbl1, tbl2, }{two object of class sstables of the same template. Only export for technical use.}

\item{footer}{[NULL] default is the joint footer of two ss_tbl}

\item{...}{args passed to the downstream method}
}
\description{
Row-binding two sstables
}
