% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkits.R
\name{ref_lv}
\alias{ref_lv}
\alias{ref_lv.factor}
\alias{ref_lv<-}
\alias{ref_lv<-.factor}
\title{Get or set the reference level}
\usage{
ref_lv(x)

\method{ref_lv}{factor}(x)

ref_lv(x) <- value

\method{ref_lv}{factor}(x) <- value
}
\arguments{
\item{x}{A factor}

\item{value}{An existing level to be set as reference}
}
\value{
A character object the represents the reference level
}
\description{
A function to quickly get or set the reference level of a factor
}
