% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkits.R
\name{simple_recode}
\alias{simple_recode}
\title{Do a simple recode via pattern for data}
\usage{
simple_recode(x, map, as = c("as_is", "numeric", "factor", "character",
  "logical"), ignore.case = FALSE, perl = TRUE, ...)
}
\arguments{
\item{x}{a vector}

\item{map}{A map.

Map can be a list in the form of recoded_text = c(to_be_recoded_text)

Or as a data frame having 2 columns named "from" and "to".}

\item{as}{A character string defining the post-recoded data type of x. Default is keeping as-is.}

\item{ignore.case, perl}{Parameters passed to gsub().}

\item{...}{Additional parameters passed to factor().}
}
\value{
A vector or length == length(x)
}
\description{
A function to re-encode x by using a map. Patterns are accepted.
}
