% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_relevel.R
\name{simple_relevel}
\alias{simple_relevel}
\alias{simple_relevel.data.frame}
\alias{simple_relevel.default}
\title{Relevel in correspondence to another factor}
\usage{
simple_relevel(x, ...)

\method{simple_relevel}{data.frame}(data, ..., by)

\method{simple_relevel}{default}(x, by, map)
}
\arguments{
\item{x}{a vector that will be relevelled}

\item{...}{Unquoted variable names that will be relevelled}

\item{data}{A data.frame}

\item{by}{a vector that the relevelling process will base on}

\item{map}{a list that links by to x in the form of list(by.1 = c(x1.1, x1.2...), by.2 = c(x2.1, x2.2...))
or a data.frame with two columns "from" and "to" serving the same purpose.}
}
\value{
A vector or a data.frame

A data.frame
}
\description{
Function to releve a variable based on a another variable of class "factor".
}
