% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_relevel.R
\name{simple_sort}
\alias{simple_sort}
\title{Sort a vector in a user-defined order}
\usage{
simple_sort(x, order = levels(as.factor(x)))
}
\arguments{
\item{x}{A vector to be sorted}

\item{order}{A manually defined order vector}
}
\value{
A vector
}
\description{
A function to sort a vector i an user-defined order
}
