% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{ss_body}
\alias{ss_body}
\title{Designation of body part for custom sstable}
\usage{
ss_body(sstable, rows)
}
\arguments{
\item{sstable}{a data frame following sstable's grammar}

\item{rows}{a numeric vector}
}
\value{
a matrix of class ss_tbl
}
\description{
This function set the designated rows as body part of a sstable
}
