% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{ss_flextable}
\alias{ss_flextable}
\title{Create summary table using flextable package.}
\usage{
ss_flextable(sstable, footer = NULL, bg = "#F2EFEE", ...)
}
\arguments{
\item{sstable}{a data frame following sstable's grammar}

\item{footer}{a character vector each of which is the footnote of the flextable}

\item{bg}{a character string that defines stripped background color of the flextable}

\item{...}{additional parameters that will be passed to ss_format if the sstable has yet to be formatted.}
}
\value{
an object of class flextable
}
\description{
This function generate a flextable from a sstable.
}
