% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{ss_format}
\alias{ss_format}
\title{Quick format a sstable in preparation for create flextable/huxtable}
\usage{
ss_format(
  sstable,
  header = c(),
  section = c(),
  body = c(),
  template = c("baseline", "survcomp", "ae"),
  .guess = TRUE
)
}
\arguments{
\item{sstable}{a data frame following sstable's grammar}

\item{header}{a numeric vector that will be passed to ss_header}

\item{section}{a numeric vector that will be passed to ss_section}

\item{body}{a numeric vector that will be passed to ss_body}

\item{template}{An accepted template for sstable: must be either 'baseline', 'survcomp', or 'ae'.

If NA, the existing template in the sstable is kept as-is. It no template built-in, will return a no-template sstable instead.}

\item{.guess}{a logical value. If TRUE, the function will try to get each row belongs to which part.}
}
\value{
a matrix of class ss_tbl
}
\description{
This function is the combination of ss_header, ss_body, and ss_section.
}
