% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{ss_header}
\alias{ss_header}
\title{Designation of header rows for custom sstable}
\usage{
ss_header(sstable, rows)
}
\arguments{
\item{sstable}{a data frame following sstable's grammar}

\item{rows}{a numeric vector}
}
\value{
a matrix of class ss_tbl
}
\description{
This function set the designated rows as header rows of a sstable
}
