% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{ss_huxtable}
\alias{ss_huxtable}
\title{Create summary table using huxtable package.}
\usage{
ss_huxtable(sstable, footer = NULL, bg = c(grey(0.95), "white"),
  border_width = 0.8, border_color = grey(0.75), ...)
}
\arguments{
\item{sstable}{a data frame following sstable's grammar}

\item{footer}{a character vector each of which is the footnote of the flextable}

\item{bg}{a character vector that defines background color of the flextable. If length(bg) >= 2, the table will have stripped background, otherwise plain.}

\item{border_width}{a number that defines huxtable border width}

\item{border_color}{a character string that defines huxtable border color}

\item{...}{additional parameters that will be passed to ss_format if the sstable has yet to be formatted.}
}
\value{
an object of class huxtable
}
\description{
This function generate a huxtable from a sstable.
}
