% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{ss_section}
\alias{ss_section}
\title{Designation of section title rows for custom sstable}
\usage{
ss_section(sstable, rows)
}
\arguments{
\item{sstable}{a data frame following sstable's grammar}

\item{rows}{a numeric vector}
}
\value{
a matrix of class ss_tbl
}
\description{
This function set the designated rows as section title rows of an sstable
}
