% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssformat.R
\name{ss_template}
\alias{ss_template}
\title{Set template for sstable}
\usage{
ss_template(sstable, template = c("baseline", "survcomp", "ae"))
}
\arguments{
\item{sstable}{a data frame following sstable's grammar}

\item{template}{An accepted template for sstable: must be either 'baseline', 'survcomp', or 'ae'.

If NA, the existing template in the sstable is kept as-is. It no template built-in, will return a no-template sstable instead.}
}
\value{
a matrix of class ss_tbl
}
\description{
This function set the template for sstable
}
