% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstable.R
\name{sstable.ae}
\alias{sstable.ae}
\title{Create an adverse event summary table}
\usage{
sstable.ae(ae_data, fullid_data, id.var, aetype.var, arm.var, digits = 0,
  test = TRUE, pdigits = 3, pcutoff = 1e-04, chisq.test = FALSE,
  correct = FALSE, simulate.p.value = FALSE, B = 2000,
  workspace = 1e+06, hybrid = FALSE, footer = NULL,
  flextable = TRUE, bg = "#F2EFEE")
}
\arguments{
\item{ae_data}{a data frame contains adverse event data.}

\item{fullid_data}{a data frame contains treatment arm data of all participants (not just those had adverse event).}

\item{id.var}{a character specifies name of study id variable (exists in both adverse event data and treatment arm data).}

\item{aetype.var}{a character specifies name of adverse event type variable (exists in adverse event data).}

\item{arm.var}{a character specifies name of treatment arm variable (exists in treatment arm data).}

\item{digits}{a number specifies number of significant digits for numeric statistics.}

\item{test}{a logical value specifies whether a statistical test will be performed to compare between treatment arms.}

\item{pdigits}{a number specifies number of significant digits for p value.}

\item{pcutoff}{a number specifies threshold value of p value to be displayed as "< pcutoff".}

\item{chisq.test}{a logical value specifies whether Chi-squared test or Fisher's exact test will be used to compare between treatment arms.}

\item{correct}{a parameter for chisq.test().}

\item{simulate.p.value}{a parameter for chisq.test() and fisher.test().}

\item{B}{a parameter for chisq.test() and fisher.test().}

\item{workspace}{a parameter for fisher.test().}

\item{hybrid}{a parameter for fisher.test().}

\item{footer}{a vector of strings to be used as footnote of table.}

\item{flextable}{a logical value specifies whether output will be a flextable-type table.}

\item{bg}{a character specifies color of the odd rows in the body of flextable-type table.}
}
\value{
a flextable-type table or a list with values/headers/footers
}
\description{
A function to create a simple adverse event summary table.
}
