% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstable.R
\name{sstable.baseline}
\alias{sstable.baseline}
\title{Create baseline table}
\usage{
sstable.baseline(
  formula,
  data,
  bycol = TRUE,
  pooledGroup = FALSE,
  keepEmptyGroup = FALSE,
  statistics = "med.IQR",
  continuous = NA,
  fullfreq = TRUE,
  digits = 1,
  test = FALSE,
  pdigits = 3,
  pcutoff = 1e-04,
  chisq.test = FALSE,
  correct = FALSE,
  simulate.p.value = FALSE,
  B = 2000,
  workspace = 1e+06,
  hybrid = FALSE,
  footer = NULL,
  flextable = FALSE,
  bg = "#F2EFEE"
)
}
\arguments{
\item{formula}{a formula specifies variables for rows, variable for column and third-dimension variable.}

\item{data}{a data frame to derive baseline table from.}

\item{bycol}{a logical value specifies whether the summary will be by column or by row.}

\item{pooledGroup}{a logical value specifies whether to pool all subgroups of column variable.}

\item{keepEmptyGroup}{a logical value specifying whether should keep empty groups}

\item{statistics}{a character specifies summary statistics for continuous row variables.}

\item{continuous}{a logical vector specifies whether each row variables is continuous or categorical.}

\item{digits}{a number specifies number of significant digits for numeric statistics.}

\item{test}{a logical value specifies whether a statistical test will be performed to compare between treatment arms.}

\item{pdigits}{a number specifies number of significant digits for p value.}

\item{pcutoff}{a number specifies threshold value of p value to be displayed as "< pcutoff".}

\item{chisq.test}{a logical value specifies whether Chi-squared test or Fisher's exact test will be used to compare between treatment arms.}

\item{correct}{a parameter for chisq.test().}

\item{simulate.p.value}{a parameter for chisq.test() and fisher.test().}

\item{B}{a parameter for chisq.test() and fisher.test().}

\item{workspace}{a parameter for fisher.test().}

\item{hybrid}{a parameter for fisher.test().}

\item{footer}{a vector of strings to be used as footnote of table.}

\item{flextable}{a logical value specifies whether output will be a flextable-type table.}

\item{bg}{a character specifies color of the odd rows in the body of flextable-type table.}
}
\value{
a flextable-type table or a list with values/headers/footers
}
\description{
A function to create a simple summary baseline table.
}
