% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstable.R
\name{sstable.survcomp.subgroup}
\alias{sstable.survcomp.subgroup}
\title{Summarize results for a survival model by treatment arm and subgroup}
\usage{
sstable.survcomp.subgroup(
  base.model,
  subgroup.model,
  data,
  compare.method = c("cox", "rmst"),
  compare.args = list(),
  digits = 2,
  pdigits = 3,
  pcutoff = 0.001,
  footer = NULL,
  flextable = TRUE,
  bg = "#F2EFEE",
  ...
)
}
\arguments{
\item{base.model}{a formula from which sub-group specific estimates are extracted (!! arm must be the first covariate in the model).}

\item{subgroup.model}{a formula of the form "~subgrouping.variable1+subgrouping.variable2" (!! subgrouping.variable must be factors and there should be nothing on the left-hand side of the formula).}

\item{data}{a data frame to fir the Cox survival model.}

\item{compare.method}{['cox'] a string, either "cox" for coxPH model or "rmst" for restricted mean survival time}

\item{compare.args}{a list of additional args for compare.methods, \n
For compare.method = 'cox', it is add.prop.haz.test [TRUE]: a logical value specifies whether a test for proportional hazards should be added,, additional args are fed directly to `survival::coxph`.
For compare.method = 'rmst', args are fed to `eventglm::rmeanglm`,
mandatory args are `time` the truncation time, default to the minimax of the observed time across all covariate specified in the model
`type`: [diff] a string, "diff" for difference in RMST, "ratio' for ratio of RMST, "lost.ratio" for ratio of restricted mean time lost
other optional args include: model.censoring, formula.censoring, ipcw.method. See `eventglm::rmeanglm` for more details.}

\item{digits}{a number specifies number of significant digits for numeric statistics.}

\item{pdigits}{a number specifies number of significant digits for p value.}

\item{pcutoff}{a number specifies threshold value of p value to be displayed as "< pcutoff".}

\item{footer}{a vector of strings to be used as footnote of table.}

\item{flextable}{a logical value specifies whether output will be a flextable-type table.}

\item{bg}{a character specifies color of the odd rows in the body of flextable-type table.}

\item{...}{arguments that are passed to sstable.survcomp.}
}
\value{
a flextable-type table or a list with values/headers/footers
}
\description{
A function to summarize results for a survival model by treatment arm (variable "arm") and subgroup.
}
\author{
This function was originally written by Marcel Wolbers. Lam Phung Khanh did some modification.
}
