% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_effect.R
\name{t.subgroup_logist_summary}
\alias{t.subgroup_logist_summary}
\title{Transpose a subgroup logist summary table}
\usage{
\method{t}{subgroup_logist_summary}(x)
}
\arguments{
\item{x}{an object of class subgroup_logist_summary}
}
\value{
If x is a simplified subgroup_logist_summary (aka a data.frame), return itself. Otherwise, a transposed list of class subgroup_logist_summary
}
\description{
A method to transpose logist summary table, rather than stratify by subgroup, it stratifies by covariables.
}
\examples{
y = sample(0:1, 1000, replace = T)
x1 = sample(1:100, 1000, replace = T)
x2 = sample(c("A", "B"), 1000, replace = T)
x3 = sample(c("C", "D", "E"), 1000, replace = T)
x4 = sample(c("F", "G"), 1000, replace = T)
fakefit = glm(y ~ x1*x3 + x2*x3 + x2*x4, family = binomial())
se = C306::subgroup_effect(~x3, fakefit)
t(se)
}
\seealso{
\link{t}
}
\author{
Trinh Dong Huu Khanh
}
