% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ajsurvplot.R
\name{tidy_competingevent}
\alias{tidy_competingevent}
\title{Tidy competing risk mstate survival fit.}
\usage{
tidy_competingevent(
  formula,
  data,
  weights,
  subset,
  na.action,
  main.event,
  competing.event,
  count,
  id,
  timefix,
  ...
)
}
\arguments{
\item{formula}{A standard model formula, with survival on the left and covariates on the right}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model}

\item{weights, subset, na.action, , count, id, timefix}{parameter passed to finegray model. See \code{\link[survival:finegray]{finegray}},}

\item{main.event, competing.event}{main and competing event to plot}

\item{...}{other parameters passed to finegray (excluding etype)}
}
\value{
a tibble
}
\description{
This complements gg_ajurvplot2, provides the intermediate data for advanced ggplot manipulation
}
