% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkits.R
\name{var_recode}
\alias{var_recode}
\title{Recode values with regex patterns and pipe support}
\usage{
var_recode(.data, ..., ignore.case = FALSE, perl = TRUE)
}
\arguments{
\item{.data}{A data frame to modify}

\item{...}{Replacement in the form of var = map. Maps will be passed to simple_recode().

See simple_recode for more details}

\item{ignore.case}{Specify whether the pattern is case-insensitive. Default is FALSE (case sensitive)}

\item{perl}{Are patterns follow the Perl-style regular expression.}
}
\value{
A data frame with recoded variables.
}
\description{
This support a robust replacement for data by providing relevant map in the form of var = map
}
\seealso{
\link{simple_recode}
}
