% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logist_summary.R
\name{logist_summary}
\alias{logist_summary}
\title{An alternative summary for a logistic model with OR, CI, and p-values}
\usage{
logist_summary(
  fit,
  method = c("lik.ratio", "wald"),
  stat_digits = 2,
  p_digits = 4,
  verbose = FALSE,
  sstable = FALSE,
  flextable = FALSE,
  ...
)
}
\arguments{
\item{fit}{a logistic regression model of class "glm"}

\item{method}{a string whose value is either
"lik.ratio" for CI and tests based on likelihood ratio statistics (preferred)
or "wald" for CI and tests based on Wald statistics}

\item{stat_digits}{Number of decimal digits for statistics}

\item{p_digits}{Number of decimal digits for the p-values}

\item{verbose}{logical value specifying whether to print out result and notation. Default is FALSE}

\item{sstable}{logical value specifying whether to return in sstable format. Default is FALSE. Set to TRUE forces verbose to FALSE}

\item{flextable}{logical value specifying whether to build flextable object. Default it FALSE. Set to TRUE forces sstable to TRUE and verbose to FALSE.
Like other sstable objects, you can still create flextable or huxtable afterwards by using \link{ss_flextable} or \link{ss_huxtable}}

\item{...}{additional parameters passed to \link{ss_flextable}}
}
\value{
A data frame of additional class "logist_summary" (if sstable == FALSE), a matrix of class c("summary_tbl", "ss_tbl") otherwise
}
\description{
a function to create summary table for glm logistic model
}
\seealso{
\link{ss_format}, \link{ss_huxtable}, \link{ss_flextable}
}
\author{
Marcel Wolbers, Lam Phung Khanh, and Trinh Dong Huu Khanh
}
