% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstable.R
\name{sstable.survcomp}
\alias{sstable.survcomp}
\title{Summarize results for a Cox survival model with the treatment arm (variable "arm") as the main covariate}
\usage{
sstable.survcomp(model, data, add.risk = TRUE,
  add.prop.haz.test = TRUE, medsum = TRUE, digits = 2, pdigits = 3,
  pcutoff = 0.001, footer = NULL, flextable = TRUE, bg = "#F2EFEE")
}
\arguments{
\item{model}{a formula which can be used to fit the Cox survival model. This formula can include other covariates than arm BUT arm must be the first covariate in the model.}

\item{data}{a data frame to fir the Cox survival model.}

\item{add.risk}{a logical value specifies whether the event probability ("absolute risk") at time "infinity" should be displayed.}

\item{add.prop.haz.test}{a logical value specifies whether a test for proportional hazards should be added.}

\item{medsum}{a logical value specifies whether median (IQR) of time to event should be described.}

\item{digits}{a number specifies number of significant digits for numeric statistics.}

\item{pdigits}{a number specifies number of significant digits for p value.}

\item{pcutoff}{a number specifies threshold value of p value to be displayed as "< pcutoff".}

\item{footer}{a vector of strings to be used as footnote of table.}

\item{flextable}{a logical value specifies whether output will be a flextable-type table.}

\item{bg}{a character specifies color of the odd rows in the body of flextable-type table.}
}
\value{
a flextable-type table or a list with values/headers/footers
}
\description{
A function to summarize results for a Cox survival model with the treatment arm (variable "arm") as the main covariate
}
\author{
This function was originally written by Marcel Wolbers. Lam Phung Khanh did some modification.
}
