% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_effect.R
\name{subgroup_effect}
\alias{subgroup_effect}
\title{Calculation of covariable effects within each value of another covariable that has interaction}
\usage{
subgroup_effect(
  formula,
  model,
  method = c("lik.ratio", "wald"),
  stat_digits = 2,
  p_digits = 4,
  sstable = FALSE,
  flextable = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{formula}{A formula in the the form of A + B ~ X, where

- X is the covariable to be concerned. X must be of type factor.
- A, B are the covarables having interaction with X whose effects will be calculated.
If . or blank. All covariables having interaction with X would be included.}

\item{model}{A fitted glm model}

\item{method}{a string whose value is either
"lik.ratio" for CI and tests based on likelihood ratio statistics (preferred)
or "wald" for CI and tests based on Wald statistics}

\item{stat_digits}{Number of decimal digits for statistics}

\item{p_digits}{Number of decimal digits for the p-values}

\item{sstable}{logical value specifying whether to return in sstable format. Default is FALSE. Set to TRUE forces verbose to FALSE}

\item{flextable}{logical value specifying whether to build flextable object. Default it FALSE. Set to TRUE forces sstable to TRUE.}

\item{simplify}{by default, if there is only one variable on each side of the formula and
no LHS is a binary, then the function will combine tables in each state of X to one.
Set to FALSE to avoid this behaviour.}
}
\value{
Under certain circumstances defined in param simplify, a flextable, an sstable,
a data.frame of class 'subgroup_logist_summary',
or a 'subgroup_logist_summary'/list of logist_summary, each represents one state of X.
}
\description{
This function calculate the OR, CI, an p.value within each value of specified covariable.
}
\author{
Trinh Dong Huu Khanh
}
